+++ lib/core/models/editor_configs/main_editor_configs.dart
@@ -29,6 +29,7 @@ class MainEditorConfigs extends ZoomConfigs {
     super.invertTrackpadDirection,
     this.transformSetup,
     this.enableCloseButton = true,
+    this.enableKeyboardShortcuts = true,
     this.enableEscapeButton = true,
     this.canZoomWhenLayerSelected = true,
     this.mobilePanInteraction = MobilePanInteraction.move,
@@ -42,6 +43,7 @@ class MainEditorConfigs extends ZoomConfigs {
       SubEditorMode.emoji,
       // SubEditorMode.sticker,
     ],
+    this.enableSubEditorPage = false,
     this.style = const MainEditorStyle(),
     this.icons = const MainEditorIcons(),
     this.widgets = const MainEditorWidgets(),
@@ -50,6 +52,12 @@ class MainEditorConfigs extends ZoomConfigs {
 
   /// Determines whether the close button is displayed on the widget.
   final bool enableCloseButton;
+
+  /// Whether keyboard shortcuts are enabled.
+  ///
+  /// When set to `true`, the editor responds to keyboard events (shortcuts).
+  /// If set to `false`, keyboard shortcuts are disabled.
+  final bool enableKeyboardShortcuts;
 
   /// Defines the configuration for pan interactions on mobile devices.
   ///
@@ -90,6 +98,9 @@ class MainEditorConfigs extends ZoomConfigs {
   /// Defines the safe area configuration for the editor.
   final EditorSafeArea safeArea;
 
+  /// Whether to use the sub-editor page without pushing a new route.
+  final bool enableSubEditorPage;
+
   /// Defines which sub-editors are available in the bottom-bar of the editor.
   ///
   /// The order of the tools in this list determines the order in the UI.
@@ -120,6 +131,7 @@ class MainEditorConfigs extends ZoomConfigs {
   /// others unchanged.
   MainEditorConfigs copyWith({
     bool? enableCloseButton,
+    bool? enableKeyboardShortcuts,
     bool? enableEscapeButton,
     MainEditorTransformSetup? transformSetup,
     MainEditorStyle? style,
@@ -138,9 +150,13 @@ class MainEditorConfigs extends ZoomConfigs {
     Curve? doubleTapZoomCurve,
     EditorSafeArea? safeArea,
     List<SubEditorMode>? tools,
+    bool? enableSubEditorPage,
   }) {
     return MainEditorConfigs(
+      enableSubEditorPage: enableSubEditorPage ?? this.enableSubEditorPage,
       enableCloseButton: enableCloseButton ?? this.enableCloseButton,
+      enableKeyboardShortcuts:
+          enableKeyboardShortcuts ?? this.enableKeyboardShortcuts,
       enableEscapeButton: enableEscapeButton ?? this.enableEscapeButton,
       transformSetup: transformSetup ?? this.transformSetup,
       style: style ?? this.style,