+++ lib/features/crop_rotate_editor/services/crop_desktop_interaction_manager.dart
@@ -77,7 +77,9 @@ class CropDesktopInteractionManager {
             _shiftDown = true;
             break;
           case 'Z':
-            if (_ctrlDown) onUndoRedo(!_shiftDown);
+            if (_ctrlDown && !HardwareKeyboard.instance.isAltPressed) {
+              onUndoRedo(!_shiftDown);
+            }
             break;
         }
       } else if (event is KeyUpEvent) {