+++ lib/features/filter_editor/filter_editor.dart
@@ -18,14 +18,14 @@ import '/shared/utils/file_constructor_utils.dart';
 import '/shared/widgets/layer/layer_stack.dart';
 import '/shared/widgets/transform/transformed_content_generator.dart';
 import 'constants/identity_matrix_constant.dart';
-import 'types/filter_matrix.dart';
 import 'utils/lerp_color_matrix_utils.dart';
-import 'widgets/filtered_widget.dart';
 
+export 'types/filter_matrix.dart';
 export 'utils/filter_generator/filter_addons.dart';
 export 'utils/filter_generator/filter_model.dart';
 export 'utils/filter_generator/filter_presets.dart';
 export 'widgets/filter_editor_item_list.dart';
+export 'widgets/filtered_widget.dart';
 
 /// The `FilterEditor` widget allows users to editing images with filters
 ///