+++ lib/features/main_editor/services/desktop_interaction_manager.dart
@@ -87,7 +87,11 @@ class DesktopInteractionManager {
     required Function onEscape,
     required Function(bool) onUndoRedo,
   }) {
-    if (!context.mounted || event is! KeyDownEvent) return false;
+    if (!context.mounted ||
+        !configs.mainEditor.enableKeyboardShortcuts ||
+        event is! KeyDownEvent) {
+      return false;
+    }
 
     final key = event.logicalKey.keyLabel;
 
@@ -124,7 +128,9 @@ class DesktopInteractionManager {
         _keyboardRotate(isLeftRotation: false, selectedLayers: selectedLayers);
         break;
       case 'Z':
-        if (_keyboard.isCtrlPressed) onUndoRedo(!_keyboard.isShiftPressed);
+        if (_keyboard.isCtrlPressed && !_keyboard.isAltPressed) {
+          onUndoRedo(!_keyboard.isShiftPressed);
+        }
         break;
     }