+++ lib/features/main_editor/widgets/main_editor_helper_lines.dart
@@ -46,7 +46,7 @@ class MainEditorHelperLines extends StatelessWidget {
   /// Configuration settings for the editor.
   final ProImageEditorConfigs configs;
 
-  static const double _strokeWidth = 1.25;
+  double get _strokeWidth => helperLines.style.strokeWidth;
   static const int _duration = 100;
 
   bool get _isLayerInRemovalZone => layerInteractionManager.hoverRemoveBtn;
@@ -69,10 +69,10 @@ class MainEditorHelperLines extends StatelessWidget {
                 final scale = viewer?.scaleFactor ?? 1;
 
                 final offset = viewer?.offset ?? Offset.zero;
-                final screenSize = sizesManager.screen;
+                final screenSize = sizesManager.editorSize;
                 final editorBodySize = sizesManager.bodySize;
 
-                if (configs.helperLines.isDisabledAtZoom && scale > 1) {
+                if (helperLines.isDisabledAtZoom && scale > 1) {
                   return const SizedBox.shrink();
                 }
 
@@ -175,7 +175,7 @@ class MainEditorHelperLines extends StatelessWidget {
 
   List<Widget> _buildLayerAlignLines(double scale, Size screenSize) {
     final editorCenter = sizesManager.bodySize / 2;
-    const halfStroke = _strokeWidth / 2;
+    final halfStroke = _strokeWidth / 2;
 
     final verticalOffset = (editorCenter.width +
             layerInteractionManager.verticalGuideOffset.dx -